Implement the checksum validation.
